refactor(server): split the /agent-gateway routes and gateway service/models into server/ai_gateway
