+++ AGENTS.md
@@ -94,7 +94,7 @@ matching path alone would not have sent you there.
 | Managed runtime | supervisor/worker convergence, install layout, runtime updates | [`specs/systems/harnesses/managed-runtime.md`](specs/systems/harnesses/managed-runtime.md) |
 | Agent auth | agent credentials, key vault, selections, `state.json` | [`specs/systems/agent_auth/README.md`](specs/systems/agent_auth/README.md)  |
 | Harnesses and models | `catalogs/**`, `scripts/agent-catalog/**`, probes, launch options, readiness | [`specs/systems/harnesses/README.md`](specs/systems/harnesses/README.md) and [`specs/systems/harnesses/launch-options.md`](specs/systems/harnesses/launch-options.md) |
-| Model gateway | `server/proliferate/server/agent_auth/{enrollment,usage_import,topups,budget,free_credits,worker,verification,migration}.py`, `server/litellm/**`, `server/proliferate/server/ai_magic/**` | [`specs/systems/ai_gateway/README.md`](specs/systems/ai_gateway/README.md) |
+| Model gateway | `server/proliferate/server/ai_gateway/**`, `server/litellm/**`, `server/proliferate/server/ai_magic/**` | [`specs/systems/ai_gateway/README.md`](specs/systems/ai_gateway/README.md) |
 | Integration gateway | `server/proliferate/server/integration_gateway/**`, `db/models/integration*.py`, `db/models/cloud/integration_approvals.py`, `db/store/integrations/**` | [`specs/systems/integration_gateway/README.md`](specs/systems/integration_gateway/README.md) |
 | GitHub | `server/proliferate/server/github/**`, `db/models/github_app.py`, `db/store/github_app.py`, `integrations/github/**` | [`specs/systems/github/README.md`](specs/systems/github/README.md) |
 | Billing | Stripe, meters, credits, plans, webhooks | [`specs/systems/billing/README.md`](specs/systems/billing/README.md) |
