+++ scripts/ownership-atlas/generate.py
@@ -100,7 +100,8 @@ OWNERSHIP: list[Region] = [
     Region("server/proliferate/server/setup", "server/setup", "identity", "keepers", "owned", "", "Server — control plane (Python)"),
     Region("server/proliferate/server/billing", "server/billing", "billing", "keepers", "owned", "", "Server — control plane (Python)"),
     Region("server/proliferate/server/support", "server/support", "support", "keepers", "owned", "", "Server — control plane (Python)"),
-    Region("server/proliferate/server/agent_auth", "server/agent_auth", "agent_auth ∧ ai_gateway", "keys", "shared", "9 files ⇒ ai_gateway; code split on the build list", "Server — control plane (Python)"),
+    Region("server/proliferate/server/agent_auth", "server/agent_auth", "agent_auth", "keys", "owned", "", "Server — control plane (Python)"),
+    Region("server/proliferate/server/ai_gateway", "server/ai_gateway", "ai_gateway", "keys", "owned", "", "Server — control plane (Python)"),
     Region("server/proliferate/server/ai_magic", "server/ai_magic", "ai_gateway", "keys", "owned", "", "Server — control plane (Python)"),
     Region("server/proliferate/server/integration_gateway", "server/integration_gateway", "integration_gateway", "keys", "owned", "", "Server — control plane (Python)"),
     Region("server/proliferate/server/github", "server/github", "github", "keys", "owned", "", "Server — control plane (Python)"),
