+++ server/proliferate/server/agent_auth/service.py
@@ -1,4 +1,4 @@
-"""Agent gateway auth services: key vault, auth selections, capabilities.
+"""Agent auth services: key vault, auth selections, state delivery, org policy.
 
 The P1 auth model (the agent-auth API surface): a titled,
 provider-less key vault plus per-(user, harness, surface) selection sources
@@ -30,18 +30,13 @@ from proliferate.db.store import agent_gateway as agent_gateway_store
 from proliferate.db.store.agent_gateway import (
     AgentApiKeyRecord,
     AgentAuthSelectionRecord,
-    AgentGatewayEnrollmentKeyRecord,
-    AgentGatewayEnrollmentRecord,
     DesiredAuthSource,
     OrgMemberRouteSelectionRecord,
-    list_active_enrollment_keys,
 )
 from proliferate.db.store.billing import list_entitlements
 from proliferate.db.store.billing_subscriptions import list_subscriptions
 from proliferate.db.store.organizations import list_organizations_for_user
 from proliferate.lib.infra.time.wall_clock import utcnow
-from proliferate.server.ai_gateway import budget
-from proliferate.server.ai_gateway.budget import get_gateway_enrollment_for_user
 from proliferate.server.agent_auth.selection_rules import (
     SelectionRuleError,
     validate_auth_selection_set,
@@ -59,7 +54,6 @@ from proliferate.server.billing.subjects import ensure_organization_billing_subj
 from proliferate.server.catalogs.service import supported_provider_config_kinds
 from proliferate.server.event_logging import log_cloud_event
 
-_ENROLLMENT_STATUS_NONE = "none"
 _MAX_TITLE_LENGTH = 255
 _MAX_SECRET_LENGTH = 4096
 # Selections persist an arbitrary harness_kind bounded only by the String(64)
@@ -487,66 +481,6 @@ async def annotate_selection_delivery(
             for record in harness_records:
                 applied[record.id] = harness_applied
     return applied
-
-
-# --------------------------------------------------------------------------- #
-# Capabilities + enrollment
-# --------------------------------------------------------------------------- #
-
-
-async def get_capabilities(
-    db: AsyncSession,
-    *,
-    user_id: UUID,
-) -> tuple[bool, str | None, str, bool]:
-    """Return (gateway_enabled, public_base_url, enrollment_status, credits_exhausted).
-
-    Status comes from the GOVERNING (never personal) enrollment, and
-    ``credits_exhausted`` negates the renderer's key-withholding predicate
-    (``is_gateway_budget_available``) — UI and render never disagree (AA-3).
-    """
-    enrollment = await get_gateway_enrollment_for_user(db, user_id)
-    return (
-        settings.agent_gateway_enabled,
-        settings.agent_gateway_litellm_public_base_url or None,
-        enrollment.sync_status if enrollment is not None else _ENROLLMENT_STATUS_NONE,
-        not await budget.is_gateway_budget_available(db, user_id),
-    )
-
-
-async def get_verification_verdicts(
-    db: AsyncSession,
-    *,
-    user_id: UUID,
-) -> list[AgentGatewayEnrollmentKeyRecord]:
-    """The per-harness gateway-enablement verdicts for the governing enrollment.
-
-    Surfaces the FR-3 verification loop's output additively on the capabilities
-    read (agent-auth.md): the enrollment-surface fallback rather than extending
-    the pinned ``state.json`` wire shape. Only keys with a recorded verdict are
-    returned; an unverified key is omitted.
-    """
-    enrollment = await get_gateway_enrollment_for_user(db, user_id)
-    if enrollment is None:
-        return []
-    keys = await list_active_enrollment_keys(db, enrollment_id=enrollment.id)
-    return [key for key in keys if key.verification_status is not None]
-
-
-async def get_enrollment(
-    db: AsyncSession,
-    *,
-    user_id: UUID,
-) -> AgentGatewayEnrollmentRecord:
-    """The governing enrollment for this user: the default org's, always."""
-    enrollment = await get_gateway_enrollment_for_user(db, user_id)
-    if enrollment is None:
-        raise CloudApiError(
-            "agent_gateway_enrollment_not_found",
-            "No agent gateway enrollment exists for this user.",
-            status_code=404,
-        )
-    return enrollment
 
 
 # --------------------------------------------------------------------------- #
