+++ server/proliferate/server/ai_gateway/MANIFEST.toml
@@ -1,0 +1,22 @@
+# System manifest (Organization Standard rule 2) — validated by
+# scripts/check_manifests.py; rules PROD-MANIFEST-001/002 in
+# lints/product/manifests.toml. allowed_importers is measured reality, not
+# permission-by-intention: the checker fails on drift in either direction.
+
+name = "ai_gateway"
+spec = "specs/systems/ai_gateway/README.md"
+owns = "Gateway enrollment, per-harness virtual keys, the LLM credit ledger and top-ups, usage import, and access-group verification: the model gateway's control plane."
+
+public_surface = [
+    "proliferate.server.ai_gateway.api",
+    "proliferate.server.ai_gateway.service",
+    "proliferate.server.ai_gateway.models",
+]
+
+allowed_importers = [
+    "accounts",
+    "agent_auth",
+    "billing",
+    "main.py",
+    "organizations",
+]
