+++ specs/systems/billing/deep-dive.md
@@ -73,7 +73,7 @@ billing_budget_limit
   enabled
 ```
 
-**N4 — every applicable enabled limit is evaluated independently; no tightest-cap masking.** A member breaches when ANY applicable enabled row breaches, not just the row with the lowest raw `cap_value` — a per-user day cap and an org month cap are different rates over different windows and neither dominates. Both compute paths (``reconciler.py`` (deleted, cull part 2)'s `_resolve_compute_limit_pause`, ``authorization.py`` (deleted in #2243, dark cloud-billing authorizer)'s `_compute_budget_cap_breach`) and the LLM import path ([`usage_import.py`](../../../server/proliferate/server/agent_auth/usage_import.py)'s `_enforce_org_llm_limits`) check every enabled row before deciding. Display-only, [`resolve_effective_limit`](../../../server/proliferate/server/billing/budget_limits.py) returns one row (the raw-tightest, possibly across different windows) and is explicitly documented as not an enforcement decision. Personal usage has no personal `billing_budget_limit` row. The full-replacement admin API accepts at most one row per user scope/kind/window — an API validation rule, not a stronger claim about nullable DB uniqueness.
+**N4 — every applicable enabled limit is evaluated independently; no tightest-cap masking.** A member breaches when ANY applicable enabled row breaches, not just the row with the lowest raw `cap_value` — a per-user day cap and an org month cap are different rates over different windows and neither dominates. Both compute paths (``reconciler.py`` (deleted, cull part 2)'s `_resolve_compute_limit_pause`, ``authorization.py`` (deleted in #2243, dark cloud-billing authorizer)'s `_compute_budget_cap_breach`) and the LLM import path ([`usage_import.py`](../../../server/proliferate/server/ai_gateway/usage_import.py)'s `_enforce_org_llm_limits`) check every enabled row before deciding. Display-only, [`resolve_effective_limit`](../../../server/proliferate/server/billing/budget_limits.py) returns one row (the raw-tightest, possibly across different windows) and is explicitly documented as not an enforcement decision. Personal usage has no personal `billing_budget_limit` row. The full-replacement admin API accepts at most one row per user scope/kind/window — an API validation rule, not a stronger claim about nullable DB uniqueness.
 
 ## Seats
 
@@ -105,7 +105,7 @@ Stripe is the money authority; every local table is a projection, not an indepen
 
 **N2 — a held/over-limit subject cannot CONTINUE; the reconciler pauses within one pass, and a stray `resumed` webhook re-pauses and closes.** The 15-minute reconciler (``reconciler.py`` (deleted, cull part 2)) lists every open segment, resolves its subject's live snapshot and any breached limit, and pauses the provider sandbox (`USAGE_SEGMENT_CLOSED_BY_QUOTA_ENFORCEMENT`) under the same per-sandbox lock the live path uses. A `created`/`resumed` provider webhook landing on an enforced spend hold is re-paused and closed by the webhook handler's own billing check (`webhooks/service.py` (deleted with the E2B webhook lane, cull part 1)) — a separate, narrower gate than the resume-path one. In the target state it resolves the payer the same way segment attribution does and evaluates budget caps as well as the hold, so an over-limit wake cannot stand until the next reconciler pass; today it checks only `active_spend_hold` on the personal subject — see [Current gaps](#current-gaps).
 
-**N3 — `limit_reached` and `exhausted` are distinct; purchasing credit never bypasses an active budget limit.** `AGENT_GATEWAY_BUDGET_STATUS_*` ([`constants/agent_gateway.py`](../../../server/proliferate/constants/agent_gateway.py)) has `ok`, `exhausted` (credit ran out), `limit_reached` (an admin cap bound, regardless of credit). `_enforce_subject_exhaustion` ([`usage_import.py`](../../../server/proliferate/server/agent_auth/usage_import.py)) never overwrites `limit_reached`; top-up reactivation ([`topups.py`](../../../server/proliferate/server/agent_auth/topups.py)) only clears `exhausted`. Only `_enforce_org_llm_limits` clears `limit_reached`, once every applicable limit passes and credit is positive.
+**N3 — `limit_reached` and `exhausted` are distinct; purchasing credit never bypasses an active budget limit.** `AGENT_GATEWAY_BUDGET_STATUS_*` ([`constants/agent_gateway.py`](../../../server/proliferate/constants/agent_gateway.py)) has `ok`, `exhausted` (credit ran out), `limit_reached` (an admin cap bound, regardless of credit). `_enforce_subject_exhaustion` ([`usage_import.py`](../../../server/proliferate/server/ai_gateway/usage_import.py)) never overwrites `limit_reached`; top-up reactivation ([`topups.py`](../../../server/proliferate/server/ai_gateway/topups.py)) only clears `exhausted`. Only `_enforce_org_llm_limits` clears `limit_reached`, once every applicable limit passes and credit is positive.
 
 **N5 — denials are never cached; a fixed subject succeeds within one enforcement cycle.** The billing snapshot and resume gate read fresh state every call. The adjacent gateway *access* cache (60s TTL, ``gateway/service.py`` (deleted, cull part 2)) bounds destroy/recreate propagation and is not a billing decision cache.
